chore(deployment): adjust examples
* migrate spring-nonpa example from JBoss to Wildfly
* migrate spring-pa example from JBoss to Wildfly
* restructure spring-pa pom to work with refactored Arquillian WebArchive 
  creation code
* update deployment examples table

related to CAM-11018
