+++ deployment/ejb-pa/README.md
@@ -6,9 +6,9 @@ This example demonstrates how to deploy an EJB process application.
 How to use it
 -----------------------------
 
-1. Checkout the project with Git
-2. Import the project into your IDE
-3. Build it with maven
-4. Deploy it to a camunda-bpm-platform distro of your own choice, that supports EJB's (Wildfly, Jboss)
-5. Wait a minute
-6. Check the console if you can find: 'This is a @Stateless Ejb component invoked from a BPMN 2.0 process'
+1. Checkout the project with Git;
+2. Import the project into your IDE;
+3. Build it with maven;
+4. Deploy it to a camunda-bpm-platform distro of your own choice, that supports EJBs (Wildfly);
+5. Wait a minute;
+6. Check the console if you can find: "This is a @Stateless Ejb component invoked from a BPMN 2.0 process".
