+++ deployment/ejb-pa/pom.xml
@@ -8,9 +8,8 @@
   <version>1.0-SNAPSHOT</version>
 
   <properties>
-    <version.camunda>7.11.0</version.camunda>
-    <version.arquillian>1.0.4.Final</version.arquillian>
-    <version.jboss-as>7.2.0.Final</version.jboss-as>
+    <version.camunda>7.12.0</version.camunda>
+    <version.java-ee-specs>3.0.1.Final</version.java-ee-specs>
     <maven.compiler.source>1.7</maven.compiler.source>
     <maven.compiler.target>1.7</maven.compiler.target>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
@@ -42,7 +41,7 @@
       <!-- Java EE 6 Specification -->
       <groupId>org.jboss.spec</groupId>
       <artifactId>jboss-javaee-web-6.0</artifactId>
-      <version>3.0.1.Final</version>
+      <version>${version.java-ee-specs}</version>
       <type>pom</type>
       <scope>provided</scope>
     </dependency>
