+++ deployment/servlet-pa/README.md
@@ -6,9 +6,9 @@ This example demonstrates how to deploy a simple servlet process application.
 How to use it
 -----------------------------
 
-1. Checkout the project with Git
-2. Import the project into your IDE
-3. Build it with maven
-4. Deploy it to a camunda-bpm-platform distro of your choice
-6. Check the console if you can find: 'Service Task 'Example ServiceTask' is invoked!'
+1. Checkout the project with Git;
+2. Import the project into your IDE;
+3. Build it with maven;
+4. Deploy it to a camunda-bpm-platform distro of your choice;
+6. Check the console if you can find: "Service Task 'Example ServiceTask' is invoked!".
