+++ deployment/servlet-pa/pom.xml
@@ -1,5 +1,6 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
+
   <groupId>org.camunda.bpm.quickstart</groupId>
   <artifactId>camunda-quickstart-servlet-pa</artifactId>
   <packaging>war</packaging>
@@ -7,9 +8,8 @@
   <version>1.0-SNAPSHOT</version>
 
   <properties>
-    <version.camunda>7.11.0</version.camunda>
-    <version.arquillian>1.0.4.Final</version.arquillian>
-    <version.jboss-as>7.2.0.Final</version.jboss-as>
+    <version.camunda>7.12.0</version.camunda>
+    <version.servlet-api.javax>3.0.1</version.servlet-api.javax>
     <maven.compiler.source>1.7</maven.compiler.source>
     <maven.compiler.target>1.7</maven.compiler.target>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
@@ -26,7 +26,7 @@
     <dependency>
       <groupId>javax.servlet</groupId>
       <artifactId>javax.servlet-api</artifactId>
-      <version>3.0.1</version>
+      <version>${version.servlet-api.javax}</version>
       <scope>provided</scope>
     </dependency>
   </dependencies>
