--- deployment/spring-jboss-non-pa/pom.xml
@@ -1,60 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-
-  <modelVersion>4.0.0</modelVersion>
-
-  <groupId>org.camunda.bpm.quickstart</groupId>
-  <artifactId>camunda-quickstart-jboss-spring-nonpa</artifactId>
-  <packaging>war</packaging>
-  <name>camunda BPM - Quickstarts - JBOSS - Spring Non Process Application</name>
-  <version>1.0-SNAPSHOT</version>
-
-  <properties>
-    <camunda.version>7.11.0</camunda.version>
-    <spring.version>4.3.24.RELEASE</spring.version>
-    <maven.compiler.source>1.7</maven.compiler.source>
-    <maven.compiler.target>1.7</maven.compiler.target>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-  </properties>
-
-  <dependencies>
-    <dependency>
-      <groupId>org.camunda.bpm</groupId>
-      <artifactId>camunda-engine</artifactId>
-      <version>${camunda.version}</version>
-      <scope>provided</scope>
-    </dependency>
-
-    <dependency>
-      <groupId>org.springframework</groupId>
-      <artifactId>spring-web</artifactId>
-      <version>${spring.version}</version>
-    </dependency>
-
-    <dependency>
-      <groupId>org.springframework</groupId>
-      <artifactId>spring-beans</artifactId>
-      <version>${spring.version}</version>
-    </dependency>
-
-  </dependencies>
-
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-war-plugin</artifactId>
-        <version>2.3</version>
-        <configuration>
-          <archive>
-            <manifestEntries>
-              <!-- write classpath module dependency to MANIFEST: this makes sure we 
-              can access the process engine classes -->
-              <Dependencies>org.camunda.bpm.camunda-engine</Dependencies>
-            </manifestEntries>
-          </archive>
-        </configuration>
-      </plugin>
-    </plugins>
-  </build>
-
-</project>
