+++ deployment/spring-servlet-pa-tomcat/pom.xml
@@ -8,7 +8,7 @@
   <version>1.0-SNAPSHOT</version>
 
   <properties>
-    <camunda.version>7.11.0</camunda.version>
+    <camunda.version>7.12.0</camunda.version>
     <spring.version>4.3.24.RELEASE</spring.version>
     <maven.compiler.source>1.7</maven.compiler.source>
     <maven.compiler.target>1.7</maven.compiler.target>
