+++ deployment/spring-servlet-pa-wildfly/src/main/java/org/camunda/bpm/example/spring/servlet/pa/ExampleBean.java
