+++ deployment/spring-servlet-pa-wildfly/src/main/resources/SpringExpressionResolvingTest.testResolveBean.bpmn20.xml
