+++ deployment/spring-servlet-pa-wildfly/src/test/resources/arquillian.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<arquillian xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+            xsi:schemaLocation="http://jboss.org/schema/arquillian http://jboss.org/schema/arquillian/arquillian-1.0.xsd">
+
+  <defaultProtocol type="Servlet 3.0" />
+  
+  <engine>
+    <property name="deploymentExportPath">target/deployments</property>
+  </engine>
+
+  <container qualifier="jboss" default="true">
+    <configuration>
+      <property name="jbossHome">target/wildfly/server/wildfly-${version.wildfly}</property>
+      <property name="javaVmArguments">-Xmx512m -XX:MaxPermSize=256m</property>
+      <property name="startupTimeoutInSeconds">300</property>
+<!--       <property name="javaVmArguments">-Xmx512m -XX:MaxPermSize=128m -Xrunjdwp:transport=dt_socket,address=8787,server=y,suspend=y</property> -->
+    </configuration>
+  </container>
+</arquillian>
